Get morphir-ir working
